pi3-histofy-backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Pydantic
python-dotenv
pillow
pytesseract
pdf2image